Provide battery and water info in get_state()

+++ embedcore/src/protocol/cyber_protocol.rs
@@ -150,7 +150,7 @@ pub struct PeripheralsState {
     pub plow: bool,
     pub led: bool,
     pub battery_voltage: f32,
-    pub water_scale: f32,
+    pub water_scale: u32,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
